+++ src/app/main-page/request/pettycash/pettycashrequest/components/VoucherPink/hooks/useVoucherPink.tsx
@@ -97,6 +97,8 @@ export const useVoucherPink = ({
     updatePettyCashVoucher,
     resetFlags,
     fetchPettyCashFunds,
+    fetchPettyCashVouchersByIdEmployee,
+    fetchPettyCashVoucherById,
     pettyCashFunds,
   } = useBillingPettyCash(
     (s) => ({
@@ -104,6 +106,8 @@ export const useVoucherPink = ({
       updatePettyCashVoucher: s.updatePettyCashVoucher,
       resetFlags: s.resetFlags,
       fetchPettyCashFunds: s.fetchPettyCashFunds,
+      fetchPettyCashVouchersByIdEmployee: s.fetchPettyCashVouchersByIdEmployee,
+      fetchPettyCashVoucherById: s.fetchPettyCashVoucherById,
       pettyCashFunds: s.pettyCashFunds,
     }),
     shallow,
@@ -208,6 +212,12 @@ export const useVoucherPink = ({
 
   useEffect(() => {
     if (!opSuccess) return;
+    if (user?.idEmployee) {
+      fetchPettyCashVouchersByIdEmployee(user.idEmployee);
+    }
+    if (isEdit && dataEdit?.id) {
+      fetchPettyCashVoucherById(dataEdit.id, true);
+    }
     ResetForm();
     showAlert({
       type: "success",
@@ -225,7 +235,18 @@ export const useVoucherPink = ({
         setOpSuccess(false);
       },
     });
-  }, [opSuccess, ResetForm, showAlert, mode, resetFlags]);
+  }, [
+    opSuccess,
+    ResetForm,
+    showAlert,
+    mode,
+    resetFlags,
+    user?.idEmployee,
+    fetchPettyCashVouchersByIdEmployee,
+    isEdit,
+    dataEdit?.id,
+    fetchPettyCashVoucherById,
+  ]);
 
   useEffect(() => {
     if (!opError) return;